test: share CPU and GPU publication scenarios
